Fix gyro and accel sensitivity parameter type
The gyro sensitivity parameter was a 16 bits variable and it clamp
the value when the 2000dps is choused. This patch change this variable
to a 32bit one, that is capable to store such kind of data size.

+++ lsm9ds1_acc_gyr.c
@@ -357,8 +357,8 @@ struct lsm9ds1_acc_gyr_status {
 	int32_t on_before_suspend;
 	int32_t use_smbus;
 
-	uint16_t sensitivity_acc;
-	uint16_t sensitivity_gyr;
+	int32_t sensitivity_acc;
+	int32_t sensitivity_gyr;
 
 	int irq1;
 	struct work_struct irq1_work;
